フォルダからの読み込み時に都度DBから取得するよう変更

+++ BmsManager/Data/BmsFile.cs
@@ -55,7 +55,7 @@ namespace BmsManager.Data
 
         public double MaxBpm { get; set; }
 
-        public int Length { get; set; }
+        public long Length { get; set; }
 
         public int Notes { get; set; }
